+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+
+
+## 1.9.3 - 2019-10-30
 ### Added
 * In dashboard, if a category is selected and the back button is pressed,
 the category is unselected